added collaborator permission button to the user edit form
