+++ src/client/components/EditUserForm/CollaboratorPermissions/CollaboratorPermissions.scss
@@ -1,0 +1,5 @@
+.edit-user__form-field-premissions {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
