+++ src/client/components/EditUserForm/CollaboratorPermissions/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CollaboratorPermissions'
