+++ src/client/components/UserList/UserList.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 
 import classNames from 'classnames'
 
-import { RoleName, User, Users, UserStatus } from '@meta/user'
+import { User, Users, UserStatus } from '@meta/user'
 
 import { useAppDispatch } from '@client/store'
 import { useAssessment, useCycle } from '@client/store/assessment'
@@ -13,7 +13,6 @@ import { UserManagementActions } from '@client/store/userManagement'
 import { useCountryIso } from '@client/hooks'
 import { useToaster } from '@client/hooks/useToaster'
 
-import CollaboratorAccessModal from '../CollaboratorAccessModal'
 import Icon from '../Icon'
 import ButtonUserListExport from './ButtonUserListExport'
 import UserInvitationInfo from './UserInvitationInfo'
@@ -44,15 +43,6 @@ const UserRow: React.FC<{ user: User; showEmail: boolean }> = ({ user, showEmail
   const cycle = useCycle()
   const currentUser = useUser()
 
-  const [modalOptions, setModalOptions] = useState<{ open: boolean }>({ open: false })
-
-  const _onEditPermissionsClick = () => {
-    setModalOptions({ open: true })
-  }
-
-  const _onEditPermissionsClose = () => {
-    setModalOptions({ open: false })
-  }
   const removeInvitation = useCallback(() => {
     if (window.confirm(i18n.t('userManagement.confirmDelete', { user: user.name })))
       dispatch(
@@ -76,21 +66,6 @@ const UserRow: React.FC<{ user: User; showEmail: boolean }> = ({ user, showEmail
       <UserRoleColumn user={user} />
       {showEmail && <UserColumn user={user} field="email" />}
       <td className="user-list__cell user-list__edit-column">
-        {user.roles[0].role === RoleName.COLLABORATOR && (
-          <>
-            <button key={1} className="btn-s btn-link" onClick={_onEditPermissionsClick} type="button">
-              {i18n.t<string>('userManagement.editPermissions')}
-            </button>
-
-            <CollaboratorAccessModal
-              open={modalOptions.open}
-              user={user}
-              headerLabel={i18n.t<string>('userManagement.editPermissions')}
-              onClose={_onEditPermissionsClose}
-            />
-          </>
-        )}
-
         {user.roles[0].invitationUuid && !user.roles[0].acceptedAt ? (
           <>
             <button
